pyproject.toml
[tool.poetry]
name = "example-project"
version = "0.1.0"
description = ""
authors = ["Tag-Epic <tagepicuwu@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pymongo = "^3.12.0"
argon2-cffi = "^20.1.0"
itsdangerous = "^2.0.1"
"snowflake.py" = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
